refactor: optimize company risk overview component
