+++ src/lib/company-risk-overview-utils.ts
@@ -1,8 +1,6 @@
-import type { CompanyRiskOverview } from '@/types/company-risk-overview'
+import type { CompanyRiskOverview } from '@/types/company-risk-overview-type'
 
-const API_BASE_URL = process.env.NODE_ENV === 'development' 
-  ? 'http://localhost:3000/api'
-  : process.env.NEXT_PUBLIC_API_URL || 'http://localhost:3000/api'
+const API_BASE_URL = 'http://localhost:3000/api'
 
 export async function fetchCompanyRiskOverview(companyId: string = 'demo-company-123'): Promise<CompanyRiskOverview> {
   try {
